+++ docusaurus-search-local/src/client/utils/__mocks__/proxiedGenerated.ts
@@ -1,10 +1,12 @@
 export let language = ["en", "zh"];
 export let removeDefaultStopWordFilter = false;
 export let removeDefaultStemmer = false;
-export const indexHash = "abc";
+export const searchIndexUrl = "search-index.json?_=abc";
 export const searchResultLimits = 8;
 export const searchResultContextMaxLength = 50;
 export const explicitSearchResultPath = false;
+export const docsPluginIdForPreferredVersion = undefined;
+export const indexDocs = true;
 
 export function __setLanguage(value: string[]): void {
   language = value;