+++ docusaurus-search-local/src/declarations.ts
@@ -11,10 +11,14 @@ declare module "*/generated.js" {
     mark: (terms: string[], options?: Record<string, unknown>) => void;
     unmark: () => void;
   }
-  export const indexHash: string | undefined;
+  export const searchIndexUrl: string;
   export const searchResultLimits: number;
   export const searchResultContextMaxLength: number;
   export const explicitSearchResultPath: boolean;
+  export const searchBarShortcut: boolean;
+  export const searchBarShortcutHint: boolean;
+  export const docsPluginIdForPreferredVersion: string;
+  export const indexDocs: boolean;
   // These below are for mocking only.
   export const __setLanguage: (value: string[]) => void;
   export const __setRemoveDefaultStopWordFilter: (value: boolean) => void;