+++ docusaurus-search-local/src/server/utils/validateOptions.ts
@@ -11,6 +11,8 @@ const isStringOrArrayOfStrings = Joi.alternatives().try(
   Joi.array().items(Joi.string())
 );
 
+const isBooleanOrString = Joi.alternatives().try(Joi.boolean(), Joi.string());
+
 const isArrayOfStringsOrRegExpsOrStringOrRegExp = Joi.alternatives().try(
   Joi.array().items(Joi.alternatives().try(Joi.string(), Joi.object().regex())),
   Joi.string(),
@@ -24,7 +26,7 @@ const schema = Joi.object<PluginOptions>({
   docsRouteBasePath: isStringOrArrayOfStrings.default(["docs"]),
   blogRouteBasePath: isStringOrArrayOfStrings.default(["blog"]),
   language: isStringOrArrayOfStrings.default(["en"]),
-  hashed: Joi.boolean().default(false),
+  hashed: isBooleanOrString.default(false),
   docsDir: isStringOrArrayOfStrings.default(["docs"]),
   blogDir: isStringOrArrayOfStrings.default(["blog"]),
   removeDefaultStopWordFilter: Joi.boolean().default(false),
@@ -34,6 +36,9 @@ const schema = Joi.object<PluginOptions>({
   searchResultContextMaxLength: Joi.number().default(50),
   explicitSearchResultPath: Joi.boolean().default(false),
   ignoreFiles: isArrayOfStringsOrRegExpsOrStringOrRegExp.default([]),
+  searchBarShortcut: Joi.boolean().default(true),
+  searchBarShortcutHint: Joi.boolean().default(true),
+  docsPluginIdForPreferredVersion: Joi.string(),
 });
 
 export function validateOptions({